Added a list of authors and contributors.
The authors were generated with the `git` command provided in the file and
cleaned up. Some of the names/emails were clearly bogus (one email was
attributed to several names, so their emails are now marked as `<?>`); some
names were placeholders or emails were using `@users.noreply.github.com`, etc.

The author (aka copyright holder) list is assumed to be DataStax due to the CLA
which included copyright assignment. It's possible that there are other
copyright holders who contributed to Titan prior to DataStax's acquisition of
Aurelius, but that would need to be verified separately.

Signed-off-by: Misha Brukman <mbrukman@google.com>