Added Travis email notifications.
Also added link to Travis status dashboard and a pointer to the mailing list
which will collect Travis status notifications.

Note: the URL is added on travis-ci.org, which is where it will be once the repo
and Travis are made public; right now, it's private so that URL will not work
unless you change `s/org/com/` to access travis-ci.com which is used for private
repos.

Signed-off-by: Misha Brukman <mbrukman@google.com>