Change module versions to 0.1.0-SNAPSHOT.
We forked Titan at the unreleased `titan11` branch where the version number was
1.1.0-SNAPSHOT, but given the amount of changes we are making and are likely to
make prior to the first release, we should inform our users that our code layout
and hence APIs are not yet stable (which a 1.x version signifies).

Once we have finalized changes, upgraded our dependencies, etc., we can upgrade
the version number to 1.0, but we cannot downgrade a release from a "stable" 1.x
down to a 0.x at any point.

Signed-off-by: Misha Brukman <mbrukman@google.com>