Fix package names and dirs in more places.
This fixes build warnings in *.java files referring to non-existent packages,
but also addresses test exclusions which will be useful in the future once we
are actually running all the tests.

Turns out, the earlier change only changed the first occurrence of
com.thinkaurelius.titan -> org.janusgraph due to the use of `'s/.../.../'` but
to change all mentions, we need to run `'s/.../.../g'`. Turns out, setting one's
editor to treat the former implicitly as the latter makes one take it for
granted.

This change was generated via the same script as earlier, but with the `g`
suffix to the `sed` command:

```bash
#!/bin/bash

grep -r -E 'com\.thinkaurelius\.titan' * \
  | sed 's/:.*$//' \
  | uniq \
  | xargs -I {} -- sed -i '' -e 's/com\.thinkaurelius\.titan/org.janusgraph/g' {}

for file in $(find . -name \*\.xml) $(find . -name \*\.properties); do
  if [ -n "$(grep com/thinkaurelius/titan "${file}")" ]; then
    sed -i '' -e 's,com/thinkaurelius/titan,org/janusgraph,g' "${file}"
  fi
done
```