Clarify differences in code and docs licensing
Per conversation with Scott Nicholas from The Linux Foundation, added
clarification regarding licensing for code (Apache 2.0) and documentation
(Creative Commons), covering both historical aspects as well as the expectation
on the new contributions going forward. This distinction in licensing is also
covered explicitly in the Project Charter document.

In addition to updating `LICENSE.txt` with copies of both licenses, also added
full copies of Apache 2.0 and CC-BY-4.0 licenses in separate files to be able to
reference them easily from the contributing guides, `pom.xml` files, the
website, and other places.

The CC-BY-4.0 license is a verbatim copy of this URL:
https://creativecommons.org/licenses/by/4.0/legalcode.txt

Signed-off-by: Misha Brukman <mbrukman@google.com>