Rename private inner class to avoid collision upon automated rename.
TitanFeatures contains an inner class TitanGraphFeatures. Both classes will
be renamed as "JanusGraphFeatures" by sed, and to avoid this we have renamed
the inner class TitanGraphFeatures to TitanGraphGeneralFeatures, producing
an outer/inner class pair of JanusGraphFeatures/JanusGraphGeneralFeatures
after the bash script to rename Titan to Janus across the repo is run.

Signed-off-by: Dylan Bethune-Waddell <dylan.bethune.waddell@mail.utoronto.ca>