Remove Faunus images and config files.
Faunus has been merged into Titan proper as of Titan 0.5 and thus, these are no
longer needed or used. There are no references to any of the images in the docs,
and the `faunus_*.txt` are no longer generated as part of the build and hence
can be removed.

In contrast, the file listing all the configuration parameters *is* still being
generated and is still referenced from the docs, and hence is not being removed.

Signed-off-by: Misha Brukman <mbrukman@google.com>