Rename interface TitanGraphTransaction to Transaction.
Run corresponding search and replace for files in which TitanGraphTransaction
is found. I do feel that the interface TitanGraphTransaction should be merged with
the interface TitanTransaction down the line. TitanGraphTransaction's influence was
limited to being extended by the interfaces TitanTransaction and TitanGraph, typing
graph or transaction variables in a handful of tests, and being imported by
org.janusgraph.graphdb.internal.InternalElement but unused therein. I tried to have
TitanGraph extend TitanTransaction and simply abolish the TitanGraphTransaction
interface, but that would have changed the API of TitanGraph. I don't want to open
a potential can of  worms orthogonal to the goal of this PR, so I'm leaving this
be and will open an issue about it.

Signed-off-by: Dylan Bethune-Waddell <dylan.bethune.waddell@mail.utoronto.ca>