Search and replace "Titan" by "JanusGraph" across repo.
Edit classnames and variables, configuration strings, and environment variables to
effectively replace occurences of "titan" with "janusgraph" throughout the repo.
Also replace references to package com[./]thinkaurelius([./]titan)? by org.janusgraph
mostly in pom.xml files and docs. Skip URLs so they can be resolved in isolation by
skipping lines that contain the string "http". I did not find any worrying cases where
barewords containing the string "http" occurred in the same line as a pending
substitution.

Also attempted to replace the aurelius-release maven profile with janusgraph-release,
and the Gremlin plugin name aurelius.titan with janusgraph.convenince pending better
candidate names for those entities. I will note that many mentions of "aurelius" are
made in the dist package that must be dealt with eventually, and the same is true of
"DataStax" in the docs.