Replace prose mentions of Titan with JanusGraph.
These cases were missed in the first pass (mostly) because we filtered out URLs
by excluding lines that contain the string "http" in our search and replace effort.
After a few grep commands it appeared safe to consider "Titan" with a space or double
quote on either side, an apostraphe on the right, or a camel case-ed suffix followed
by space or a backtick to be prose and desirable to replace by "JanusGraph". That is
what is reflected in the regex used by sed.

Signed-off-by: Dylan Bethune-Waddell <dylan.bethune.waddell@mail.utoronto.ca>