Renamed strings in documentation.
The usual renaming changes took place:
* s/Titan/JanusGraph/
* s/TitanManagement/JanusGraphManagement/
* s/titangraph/janusgraph/
* s/titan/janus/ (except where it would double the word `graph`)
* etc.

Also updated the Gremlin plugin name, module names, file names, environment
variables, etc. Removed content or historical references where no longer
applicable.

Verified that the docs build works via:

```bash
$ mvn clean package -DskipTests=true
```

and manually inspecting the output in a browser.

The most complicated part was debugging the build failures. It turns out that
AsciiDoc is sensitive to the number of characters used to underline a header
(which signifies its level in the hierarchy). Getting it wrong breaks the build
in the most convoluted way (i.e., the error message is completely unhelpful and
will not lead you to fix the error).

Specifically, before we had headings of the form:

```
Titan Documentation
===================
```

but note that by replacing `s/Titan/JanusGraph/`, we are increasing the title by
5 characters, while leaving the underline the same length:

```
JanusGraph Documentation
===================
```

and this causes a build error such as:

```
[...]/janusgraph-dist/target/docs/docbook/index.xml; lineNumber: 3567; columnNumber: 485; name expected (found "<") -> [Help 1]
```

and others like it. Also, it turns out that there are various other header
markers, such as:

```
Title
=====

Section 1
---------

Section 2
~~~~~~~~~

Section 3
^^^^^^^^^

Section 4
+++++++++
```

It turns out, that 3 character or less underflow on the underline characters is
OK, but since we have 5, each of those section titles which are not fixed will
break the build in a subtle way. Sometimes, it even lets you complete the build,
but the content does not entirely appear in the output HTML pages.

The way to track this down was to use `grep` with context lines:

```bash
$ grep -B 1 -- '=====' *.txt
```

and inspecting the line lengths of the section title name and underline marker.

The [AsciiDoc
documentation](http://asciidoctor.org/docs/asciidoc-recommended-practices/#section-titles)
describes this as the "Setext" style section title markup and strongly
discourages it for this exact reason (emphasis in the original):

> I _strongly_ recommend against using the Setext style for section titles.
> Keeping the length of the underline in sync with the title length is an
> unnecessary task and time waster. A more substantial reason is that the
> mapping between character and heading level is *very* difficult to remember and
> not obvious to new AsciiDoc adopters. All in all, using Setext section titles
> is a _bad practice._ *Don’t use them!*

Signed-off-by: Misha Brukman <mbrukman@google.com>