Avoid timeouts due to no output with `travis_wait`
The last line of the output of a recently-failed Travis job:
https://travis-ci.org/JanusGraph/janusgraph/jobs/195761554
tells us that our build is being terminated due to no output for 10 minutes.

The output also tells us to look at the following documentation:
https://docs.travis-ci.com/user/common-build-problems/#Build-times-out-because-no-output-was-received
which suggests that using `travis_wait` command for silent long-running commands
will print out a line of output every minute for 20 minutes, such that Travis
does not automatically end the build if there's no output for 10 minutes.

If we need more time, per the docs, we can extend it by adding a number to the
command, e.g.,

```
travis_wait 30 mvn clean package [...]
```

Signed-off-by: Misha Brukman <mbrukman@google.com>