Enable testing of all modules on Travis CI
This change enables Travis CI to test all modules, even those we know to be
failing, but annotations the expected failures explicitly, such that their
failures will not fail the overall build.

As a result, we will be able to track improvements to the tests to allows us to
remove them from the `allowed_failures` list in the future, once fixes have been
made.