Clarify how to subscribe to mailing lists
We are getting more and more emails with the text "subscribe" so let's clarify
the process for subscribing to the mailing lists as it appears it's non-obvious.

The new text provides a click-to-subscribe process; it also makes the email
address not a link so that it's somewhat slightly harder for spammers to
discover the email list, and new users won't be tempted to click on the email
address to send the text "subscribe" to it.

Signed-off-by: Misha Brukman <mbrukman@google.com>