Modifications to HBase storage to support Google Cloud Bigtable.
1. Moved getRegionLocations to the ConnectionMask interface. Implemented
in 0_98 implementation using the existing method, in 1_0 implementation
uses connection.getRegionLocator method instead.

2. Cache the returned value of HBaseStorageManager.getDeployement so
that subsequent calls are fast.  This is called a lot during queries,
and this requires a backend call, so is expensive. Using Bigtable
without this cache you quickly hit the quota limits for ListTables.

Tested by running full test suite in the janusgraph-hbase-parent folder.

Change ConnectionMask interface to use 1.0 style HRegionLocations as the return type.  Update normalizeKeyBounds to operate on this list.

Add documentation for connecting to Bigtable backend