Implemented a new record reader to extract JanusGraph data from Cassandra-3.
Summary:
The new reader and key iterator are inspired by Cassandra-3's CqlRecordReader.
Preparing for a PR upstream. Tested thoroughly on local C-3 instance on OLAP
path via SparkGraphComputer.

Note that the way I am plugging in
the Reader is by keeping the Cassandra-2 path unaffected. Only when the system
property "use.cassandra.3" is defined, this code path gets exercised.
To embed JG as a library in a JVM, I resorted to using a system property
instead of a different InputFormat. So, right now, we have a single InputFormat
for both Cassandra-2 and Cassandra-3, although it behaves differently based
on the backend.

Tested locally on Cassandra-3 per instructions from Ted Wilmes (see the
issue-172 page).