Check if vertex tx open on cache removal
Since the Guava cache [processes its removals
asynchronously](https://github.com/google/guava):

"""
Caches built with CacheBuilder do not perform cleanup and evict values
"automatically," or instantly after a value expires, or anything of the
sort. Instead, it performs small amounts of maintenance during write
operations, or during occasional read operations if writes are rare.
The reason for this is as follows: if we wanted to perform Cache
maintenance continuously, we would need to create a thread, and its
operations would be competing with user operations for shared locks.
Additionally, some environments restrict the creation of threads, which
would make CacheBuilder unusable in that environment.
"""

there is a chance that we reach the onRemoval code _after_ the vertex
being removed due to cache eviction asynchronous processing has indeed
_already been removed and processed_ by user interaction. In this
scenario, the transaction is not open, and more importantly, this
denotes that the vertex processing is done and we do not need to add the
vertex to the volatileVertices map.

Note that the Vertex's tx() method returns this transaction if it
exists, or _the next_ transaction otherwise. We need to ensure we are
checking that _this_ transaction is open.

Issues: #503