Assume empty list when query returns no result
This changeset fixes a scylla-cassandra compatibility issue in thrift.  Cassandra returns an empty list for a slice it doesn't find while ScyllaDB returns a null pointer.  In either case, we don't want to return a null object for a vertex.  We should return an empty list when no vertex ID is returned.

Issue: #557
Signed-off-by: David Pitera <dpitera@us.ibm.com>