Add Google Cloud Bigtable module.
This makes it very easy to build JanusGraph with Bigtable included, or link in
JanusGraph support with Bigtable directly into an application, by providing an
artifact that has all the right libraries.

Updated the docs to remove the directions to manually add the necessary
dependencies, since they will be included in `janusgraph-all` after this
change

This commit resolves issue #587.