Sync EvictionTrigger's openInstances on schedule
Without this change, any job to REGISTER an index requires
acknowledgement from every server on which the graph was open at the
time of index creation, however the index state change would never
fulfill should, for example, the DBA removed a JG node from the cluster.

Note that we do not make the EvictionTrigger take into account newly
added JG nodes because these nodes would never receive the message that
the index was is being registered and thus could never send an
acknowledgement message.

Issues: #716


Create ConfigOption REPLACE_INSTANCE_IF_EXISTS

If a user supplies their instance name (or suffix) at graph open time,
as well as use this new configuration option with a value of true, then
a graph trying to open with the same instance-id as one previously
opened will still open.

Issues: #716


ConfigOption to set graph-instance to hostname

This is useful if you want to track JG nodes by their hostnames, with
the option of appending suffix values to the end of the hostname,
whereas previously the suffix option is appended to a hex encoded local address.

Issues: #716