Rename exceptions with empty catch, use try-with-resources, misc.
This change mainly renames all exceptions that are caught with
empty catch blocks. It also fixes some throws from finally
blocks with try-with-resources and fixes some straggling issues
like some lambda method references that I missed in earlier
refactorings and various updates to javadocs.

Addresses #796 in part and #781 in part