Avoid unnecessary reindex operations #1190
This avoids the need to perform a reindex operation for cases where an
index is created for already existing property keys but with a
constraint to a label that is created in the same transaction as the
index. In those cases, there cannot be any existing data in the graph to
which the index applies as the label was not present before. So, a
reindex operation is not necessary.