Add docs on how to connect to JanusGraph Server [skip ci]
The new part in the docs simply shows examples for frequently used
languages on how to connect to JanusGraph Server and points to the lists
of available TinkerPop drivers and languages. So, users can also just
pick one from there in case they are using a language that is not
covered by these docs.

Fixes #205