Fixing issue 1333
Created a new mutex to lock during the updateLifeCycle method. Not sure what the origional intention of having this method be synchronized was, but this should ensure the behavior is the same without the risk of deadlock.

Also marked lifecycle as volatile since it can be updated and read by different threads.

Signed-off-by: Mike Beaubien <mbbeaubi@umich.edu>