Use new Credential DSL for authentication #1364
TinkerPop deprecated the old CredentialGraph and it will be removed
completely in version 3.4.0. Instead, the new Credential traversal DSL
should be used.
So, this is a precondition for the upgrade to TinkerPop 3.4.0.

Since the tests contained a lot of duplicated code, extensive use of
mocking and were relatively long in general, it was hard to understand
them. I therefore refactored them completely which should make the tests
easier to understand.
During the refactoring of the tests I also noticed some minor issues in
the code which I also fixed.