Fix relation removal in other transaction #1466
The vertex equality check in AbstractTypedRelation was used to check
whether the vertex was already loaded in the current transaction. This
check only works if both objects are actually the same, e.g., with
`v == v.it()`. `v.eqauals(v.it())` doesn't work as it is always true
because equals() only checks whether the ids are equal. So, this check
cannot be used to check whether the vertex was loaded in the
transaction.

This reverts PR #1118 basically which introduced this problem.
The StackOverflowError that was fixed by PR #1118 is now instead fixed
by removing the unnecessary overload of longId() in CacheVertexProperty.
This overload produced the stack overflow basically like this:

1. The id of a vertex property is retrieved. This calls
    CacheVertexProperty.longId().
2. To get the id, it() was called to load the property in the context of
    the transaction.
3. it() needs the id to load the property. -> return to step 1.

Without the CacheVertexProperty.longId() overload, this circle doesn't
occur as the id is directly returned instead of calling it().

The TinkerPop tests are now also executed on Travis, at least for the
in-memory backend. This hopefully prevents such problems in the future
as the TinkerPop tests are quite extensive.