Upgrade to TinkerPop 3.4.1 #1364
* TinkerPop text predicates added in TinkerPop 3.4.0 are supported, but
they currently cannot use indices. We can improve that in the future.
* FulgoraGraphComputer needed some fixes to get all tests passing and
some TinkerPop tests needed to be ignored. Some simply don't work for
JanusGraph as they make assumptions about the ordering of elements
returned which is (more or less) random. The newly added vertex programs
ShortestPathVertexProgram and ConnectedComponentVertexProgram also seem
to be problematic for FulgoraGraphComputer which required some special
treatment (see changes in VertexProgramScanJob and TINKERPOP-2187).
* The way HBase is started for the TinkerPop tests needed to be fixed as
TinkerPop apparently ignores the JUnit BeforeAll annotations that were
used before to start HBase. The HBase tests now start HBase the same way
we already do it for the Cassandra/CQL tests.
* TinkerPop tests are now executed in their own Travis job as the time
limit was reached when the tests where executed together with the normal
tests.

This doesn't remove usages of deprecated code or mentions in the docs
of such deprecated code which includes usages of Order.incr and
Order.decr that were replaced by Order.asc and Order.desc and the way
traversal sources are constructed which should use the newly introduced
AnonymousTraversalSource.
These issues should be taken care of in a follow-up contribution.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>