Fix testNestedWrites testcase in JanusGraphIndexTest
The test case builds a mixed index using TEXT mapping, but uses
equality condition in query (which does not use this mixed
index). This makes one assertion does not work as expected, and
as a consequence, passes the test by chance. This commit fixes the
mixed index usage and the assertion to uncover the real problem.

Related to #2176

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>