Make iterators auto-closeable and fix SubqueryIterator close bug
Changes:

1. Make a few iterators auto-closeable. Now traversal.close() can
be delegated to SubqueryIterator.
2. Extract some logic from MultiDistinctOrderedIterator and use new
MultiDistinctUnorderedIterator when orders are not given. This
makes the code cleaner and reduces the overhead of creating &
maintaining complex data structures.
3. Make a few iterators extend CloseableAbstractIterator which in
turn leverages Guava Iterator's computeNext method. Now, these
iterators don't need to explicitly write hasNext and next.
4. Fix SubqueryIterator bug so that profiler is correctly stopped
and results can be cached appropriately.

This commit does not:

1. Make all iterators auto-closeable. It only refactors iterators on
the traversal -> SubqueryIterator chain so that traversal.close()
can be delegated.
2. Refactor all iterators using computeNext method. Some iterators have
complex internal logic thus I would like to leave them as they are
in this commit. Generally, all iterators using nextInternal method
could be rewritten using computeNext which is much cleaner and
avoids eager-computing (fetches the first result in the constructor).

Fixes #2217

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>