Fix cached relation update hashcode change bug
When a CacheEdge or a CacheVertexProperty is updated, JanusGraph
creates a new edge/property with auto-generated longId, and then overrides
it with the old id. Since AddedRelationsContainer implementation
relies on the longId as part of hash key, creating a new edge/property
with new longId can lead to a commit exception. This bug appears
when a vertex property or an edge is updated and then deleted in the
same transaction.

Fixes #2138

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>