Bug fixes: relations cannot be removed after update
This commit includes fixes for three bugs:

1) Caused by the optimization in StandardJanusGraphTx::addProperty, which saves a database
read (and delete) operation for vertex property upsert. If a vertex property (or the vertex
as a whole) is removed after the property is modified in the same transaction, then the old
value in database will not be deleted after the transaction commits. Reported by #1981.

2) Caused by CacheVertexProperty::remove method which does nothing if it finds the given
vertex property is already in the deletedRelations map. If the user deletes a vertex property
after they update a nested property of that vertex property in the same transaction, then
the old value will be removed from database but the new value will be saved into database
after the transaction commits.

3) Caused by CacheEdge::remove method which does nothing if it finds the given edge is
already in the deletedRelations map. If the user deletes an edge after they update a property
of that edge in the same transaction, then the old value will be removed from database but the
new value will be saved into database after the transaction commits.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>