Fix nested AND/OR profiler group issue
Currently, when there is more than one backend query, annotations
(condition, orders, isFitted, isOrdered, query, index) of the last
one overwrites previous annotations in the profiling result.

This commit fixes this problem by avoiding flattening AND/OR query
groups. The drawback of this commit is those groups (AND-query, OR-query)
are not timed due to technical difficulty, which is left as future
work.

Fixes #2285

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>