Add CacheVertex::refresh method to allow clearing vertex cache
In some use cases, users might want to read from the underlying storage
backend again in the mid of a transaction, bypassing the vertex cache.
This commit adds a refresh method to CacheVertex which user could call
to clear the cache.

Related to https://lists.lfaidata.foundation/g/janusgraph-users/topic/81050716

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>