Fix profiling when limit adjusting is enabled
When LimitAdjustingIterator is in effect, multiple backend queries with
different limits (e.g. 1000, 2000, 4000, ...) are fired progressively.
They are, however, not all profiled. Currently, only the last backend
query is profiled because of a bug. This commit fixes it and now users
can see timing of each backend query to better understand whether they
need smart-limit or not.

This commit also fixes a minor thing with profiler output, such that when
QUERY.NO_LIMIT is used or no limit is given, we don't display Integer.MAX_VALUE
in the profile output.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>