CQLKeyColumnValueStore: Avoid fetching WRITETIME and/or TTL if no need
When meta.*.timestamps and/or meta.*.ttl are disabled (which are by default),
write timestamps and TTL info will not be retrieved. CQLKeyColumnValueStore,
however, retrieves them anyways, and simply discard them. This commit makes
CQL driver fetches them only when needed.

It also enhances relevant documentation so that users know how to make use of
these meta data if needed.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>