Refactor JanusGraphSchemaVertex to return List instead of Iterable
getRelated is always computed as List so there is no sense to return Iterable instead of List.
In case the implementation of getRelated is changed to return Iterable then we can change the
dependant logic back to use Iterable instead of List.
Right now using List instead of Iterable allows us to create ArrayList in some places with correct size
instead of LinkedList which generally takes more space.

Signed-off-by: Oleksandr Porunov <alexandr.porunov@gmail.com>
Signed-off-by: Oleksandr Porunov <alex@mapped.com>