Abort transaction if mixed index schema gets updated concurrently
An update to the mixed index would cause open transactions to fail and throw
NullPointerException if they happen to touch the same index. #2688 fixes this
by making schema changes transparent to all transactions. The fix is, however,
incomplete in the sense that transaction now still fails but fails silently.

Generally, there are two solutions:

1) Do not attempt to fix this issue, but provide a better error message.
2) Continue the fix in #2688 and make sure schema caches are cleared.

This commit reverts #2688 and provides a better error message when aborting
the transaction. Note that even if we choose 2) later, this commit still
remains valuable, because schema updates are propagated asynchronously
among JanusGraph instances.

This commit does not judge whether we should do 2) or not. It reverts #2688
simply because the fix is incomplete and can possibly cause more trouble than it
attempts to fix.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>