Change LOCK_WAIT configuration from GLOBAL_OFFLINE to MASKABLE
LOCK_WAIT is used in JanusGraph locking protocol to determine how long the system should wait
for a lock application to be acked by the storage backend (see ConsistentKeyLocker::writeSingleLock).
It is also the time waited at the end of all lock applications before verifying that
lock applications were successful (see ConsistentKeyLocker::checkSingleLock).

The recommended value is a small multiple of the average consistent write time, thus typically a
small number, e.g. 100 milliseconds. However, in some use cases, the consistent write time can
be much larger. For example, there can be a significant latency for developers to launch a JanusGraph
instance locally, connecting to the storage backend located in another country. In this case, a short LOCK_WAIT
would lead to timeouts and thus transactions would fail. Thus, it would be useful if this LOCK_WAIT
is maskable. Despite this legitimate usecase, this commit adds a notice to remind users of potential
risks when using different LOCK_WAIT values across JanusGraph instances.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>