Explicitly add commons-configuration2 dependency
org.apache.commons:commons-configuration2 is used widely in JanusGraph but it is not
explicitly included in pom.xml, which causes coexistence of two different versions,
one introduced by tinkerpop and the other introduced by solr-test-framework. Although
the released jars only include the tinkerpop version, which is correct, it is better to
explicitly define this dependency in pom.xml.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>