Implement our own GraphBinary Geoshape serialization
This makes it much easier to support GraphBinary outside of Java. It
also reduces the payload size by a factor of 3 to 4 (taken from the
examples used in the tests):

| Geoshape type      | Payload size (B) with previous serialization | Payload size (in B) with new serialization | Relation (new / old) |
|--------------------|----------------------------------------------|--------------------------------------------|----------------------|
| Point              | 73                                           | 22                                         | 30.1%                |
| Circle             | 105                                          | 30                                         | 28.6%                |
| Polygon            | 373                                          | 90                                         | 24.1%                |
| GeometryCollection | 633                                          | 164                                        | 25.9%                |

It is unfortunately not possible to add the Geoshape types also directly
as GraphBinary types which would simplify the serialization and also
reduce the payload size even more as we wouldn't have to serialize the
Geoshape GraphBinary type together with the Geoshape type code. This
however doesn't work with how Geoshapes are implemented right now in
Java as they are all represented by a single class `Geoshape`. We would
need to split this class into different classes for the Geoshape types
which seems out of scope for now as we just want to improve the
GraphBinary serialization here.

Fixes #2856

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>