Add distribution test for SparkGraphComputer traversal
We have SparkGraphComputer tests in AbstractInputFormatIT, but not in distribution tests.
PR #2756 is an example which shows that dependency issues could damage the functionality even
if AbstractInputFormatIT passes. This commit aims to test basic SparkGraphComputer traversal
which leverages local Spark instance in gremlin console.

The new test fails without the fix in #2756.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>