Refactor JanusGraphManagement.IndexJobFuture
JanusGraphManagement.IndexJobFuture is used in many places, not limited to
index-related OLAP jobs. The current name is confusing and misleading.

This commit renames this interface and moves it out
of JanusGraphManagement class. This commit also moves its implementations
out of parent classes, and adds respective unit tests. No functionality
change is involved.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>