Fix NullPointerException caused by transaction release
Fixes #2898. Due to a race condition, vertexCache could be cleared before
it is accessed. To avoid NPEs, we cannot simply assign null to vertexCache
to release the memory.

This commit also surpresses https://lists.lfaidata.foundation/g/janusgraph-users/topic/potential_transaction_issue/85970858
by using a dummy empty vertexCache after transaction release. The root cause
for this bug is unknown, but likely it is due to another race condition.
To keep the behavior consistent with v0.5 and before, this commit surpresses
this issue and writes a warning log whenever the issue happens.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>