Ensure Jetty versions are aligned
We're not directly using Jetty but a few dependencies have a dependency
on it which is why it's in the pom.xml so we can prevent dependency
conflicts.
We currently seem to use the same Jetty version as Solr which is also
the newest Jetty version used by any of our dependencies right now.

Using a variable for the Jetty version also enables Dependabot to update
the version for all three Jetty dependencies together.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>