Disable StaticArrayEntryListBenchmark and set FailOnError
Currently, StaticArrayEntryListBenchmark causes OOM but
benchmark CI job does not exit. This commit sets up
JMH's shouldFailOnError option so that the CI job will
fail in case of any error. Moreover, this commit excludes
StaticArrayEntryListBenchmark from the standard benchmark
suite because it only tests a low-level API that is rarely
changed and thus does not have much value in CI.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>