Update test dependencies
This updates JUnit, TestContainers, Lombok, and Mockito together as it
wasn't possible to only update TestContainers alone since it also
depends on JUnit, see PR #2972.
The same applies to Mockito.

These changes were necessary to update Mockito to v4:

Import of `org.mockito.Matchers` needed to be changed to
`org.mockito.ArgumentMatchers`. `Matchers` was already only an alias
for `ArgumentMatchers` before as described in the release notes:
https://github.com/mockito/mockito/releases/tag/v4.0.0

`any()` is the replacement for `anyObject()`:
https://github.com/mockito/mockito/pull/1931

Calling `property()` on a mocked `StandardVertex` resulted in a
`NullPointer` exception, but we can also simply use a `DetachedProperty`
as all accesses to that property on the vertex are also already mocked.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>