Fix mixed index query for neq condition
If a mixed index consists of multiple properties, then an element can be
indexed as long as it contains one of these properties. In this case, a
query like has(keyA, neq(val)) could match an element without keyA
as long as it is indexed by this mixed index. This commit fixes this bug
by adding an existence condition, i.e. matching an element only if it
has the given key present.

Fixes #3164

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>