Fix the issue [#3344]
JanusGraphVertexDeserializer parses the data, create TinkerVertex, which was transformed to StarVertex by HadoopRecordReader, and finally saved in VertexWritable. This process is like "parse data -> create TinkerVertex -> create StarVertex", which can be simplified to be "parse data -> create StarVertex".

This optimization removes the unnecessary step, and reduces the memory footprint when deserializing a huge graph.

I composed a simple testcase which calls the JanusGraphVertexDeserializer.readHadoopVertex() and calls "new VertexWritable(result)" to simulate the above process, I used "jstat -gc $PID 1s" to monitoring the impact on GC. This test has run for 3 times on the version before and after the fix.

We can observe that the total GCT was reduced from 1.778 to 1.457. There are ~22% improvement.

Before the fix:
--------------
|  | GCT (GC Time)
--------------
|1 | 1.708   |
--------------
|2 | 1.885   |
--------------
|3 | 1.742   |
--------------

After the fix:
--------------
|  | GCT (GC Time)
--------------
|1 | 1.563   |
--------------
|2 | 1.394   |
--------------
|3 | 1.415   |
--------------

Signed-off-by: Hongjiang Zhang <timerzhj@hotmail.com>
(cherry picked from commit a47acfbb46bdb9689681c8a9da3b3ef109603d22)

# Conflicts:
#	janusgraph-hadoop/src/main/java/org/janusgraph/hadoop/formats/util/JanusGraphVertexDeserializer.java