Document new backporting strategy
Fixes #3404

Unfortunately, adding the backport labels is a manual process right now
that we have to do as committers because:

* Labels cannot be added already by a pull request template:
https://github.com/community/community/discussions/12815
* Contributors cannot add them themselves as they lack the required
permissions to add labels.

We can however look into adding some action that automatically adds
labels to all PRs: https://github.com/actions/labeler
Then we would only have to remove the label again on PRs that should not
be backported (as that should be fewer than the PRs that we want to
backport).

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>