Let required checks pass for doc PRs
We want to use required checks for PRs that prevent merging of PRs if
these checks didn't pass. This also makes it possible to use
auto-merging.
However, we don't want to build and test PRs that for example only
change docs, but such PRs still need to pass the required checks.
This forces us to create duplicates of workflows with required checks
that are executed only for those PRs (where no code is changed). These
duplicate workflows will immediately turn green.
The workflows simply need to have the same name.

This approach is recommended in the GitHub docs:
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks

I simply created duplicates for all workflows that are executed on PRs,
even though we don't use most of them for required checks. This allows
us to add more checks as required if we want to without having to change
the workflows again.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>
(cherry picked from commit 10e5ef58100459bf93f922402e939b514e178d0a)