Fix potential wrong exceptions during multi-query properties or vertex steps
Before this commit the root cause exception of a failed preFetch logic could be lost and the NPE exception would be thrown instead. This commit changes logic to throw the root cause exception instead. Also, in case of interrupted exception it now attaches the root cause exception to the new thrown exception.

This commit also refactors preFetch optimization logic to decrease duplicate code of preFetch logic in `JanusGraphPropertiesStep` and `JanusGraphVertexStep`.

Signed-off-by: Oleksandr Porunov <alexandr.porunov@gmail.com>