Use Reload4j as the default logging implementation
Log4j isn't available anymore in the lib directory, but Reload4j is.
Since Reload4j is a drop-in replacement for Log4j which fixes security
issues of Log4j, it makes sense to switch to that.
This should provide an elegant, non-breaking way, to avoid the security
issues of Log4j.
We only need to do this for `v0.6` as `master` already uses Log4j2.

I verified manually that the expected logs appear again with
`bin/janusgraph-server.sh` and that the Log4j config file is used by
Reload4j.

Fixes #3681

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>