[Doc] Rename 'Interactions with JanusGraph' section to 'JanusGraph Basics'
The JanusGraph Basics section was split into a few sections in v0.6, one of which was
'Interactions with JanusGraph'. This section contains 3 things: transactions, predicates,
and drivers. It's hard to find a common topic among them, and 'interactions' is a very vague
word that is not straight-forward to understand. This renames the section back to 'JanusGraph
Basics', because transactions, predicates, and drivers are the necessary basic staff one has
to learn about in order to use JanusGraph.

Signed-off-by: Boxuan Li <liboxuan@connect.hku.hk>