Add missing JUnit annotations to Berkeley tests
Some tests weren't executed prior to the JUnit update because they
lacked the `@Test` annotation. I also added the `@Disabled` annotation
to tests which were overriden with an empty test to show the intent for
this.
One test is failing with the added `@Test` annotation, but it was
already failing before, but just wasn't executed. So, this now
explicitly ignores the test, but we should of course still find out what
the underlying problem is: #4153.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>
(cherry picked from commit acac89380b76acbc7f3cb5dae50114d3af7331a6)