Mark test as flaky for all backends
We only annotated the test for CQL as flaky, but
it's also flaky for other backends. contributor
@kptfh already commented on the issue that it's
also flaky for the Aerospike backend:
https://github.com/JanusGraph/janusgraph/issues/1498#issuecomment-475640179

and Natalia recently commented on Discord that the test is flaky for
in-memory.

So, we should just move the annotation to the test itself to be not
specific for any backend.
Since the test is copied to `JanusGraphCustomIdTest`, we also need the
annotation there (and this is also the place where Natalia hit the flaky
test).

Issue: #1498 (not fixed by this)

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>
(cherry picked from commit e95578388919af6749b6a5e81f850de08c846b2d)